feat: add explorer api
